Rename inline_labels.py into inlining.py
