+++ inline_labels/__init__.py
@@ -1,4 +1,4 @@
 __all__ = ["add_inline_labels"]
-__version__ = "0.2.1.33"
+__version__ = "0.2.1.34"
 
-from inline_labels.inline_labels import add_inline_labels
+from inline_labels.inlining import add_inline_labels
